Prevent effect from running on every render
Awesome project! Thanks for building and demploying this. :)

Including `START_DATE` in the `useEffect` dependency array causes the
effect to run on every render, since `START_DATE` is being reinitialized
every time the component rerenders.

That means that right now, every 100 milliseconds the value updates, the
interval is cleared, the effect runs, creates a new interval, etc.

An empty dependency array tells react to run it exactly once, which is
all that you need.